test: move broken test to use private constructor
